moved coiltypes
