+++ src/components/ComplexWire/index.tsx
@@ -5,6 +5,7 @@ import { FormattedMessage } from 'umi-plugin-react/locale';
 import SingleWire from '@/components/SingleWire';
 import { Path } from '@/models/coil';
 import ImageWebp from '../ImageWebp';
+import types from './coilTypes';
 
 const { Option } = Select;
 
@@ -21,73 +22,6 @@ export interface WireComponentProps {
   onSetWire: (path: Path[], wire: Wire) => void;
   onDeleteWire: (path: Path[]) => void;
 }
-
-const types: { name: string; src: any; proOnly?: boolean }[] = [
-  {
-    name: WireType[WireType.NORMAL],
-    src: require('@/assets/coil_type_normal.webp'),
-  },
-  {
-    name: WireType[WireType.PARALLEL],
-    src: require('@/assets/coil_type_parallel.webp'),
-  },
-  {
-    name: WireType[WireType.TWISTED],
-    src: require('@/assets/coil_type_twisted.webp'),
-  },
-  {
-    name: WireType[WireType.CLAPTON],
-    src: require('@/assets/coil_type_clapton.webp'),
-  },
-  {
-    name: WireType[WireType.RIBBON],
-    src: require('@/assets/coil_type_ribbon.webp'),
-  },
-  {
-    name: WireType[WireType.FUSED_CLAPTON],
-    src: require('@/assets/coil_type_fused_clapton.webp'),
-  },
-  {
-    name: WireType[WireType.ALIEN_CLAPTON],
-    src: require('@/assets/coil_type_alien_clapton.webp'),
-    proOnly: true,
-  },
-  {
-    name: WireType[WireType.TIGER],
-    src: require('@/assets/coil_type_tiger.webp'),
-    proOnly: true,
-  },
-  {
-    name: WireType[WireType.STAPLE],
-    src: require('@/assets/coil_type_staple.webp'),
-    proOnly: true,
-  },
-  {
-    name: WireType[WireType.STAGGERED_CLAPTON],
-    src: require('@/assets/coil_type_staggered_clapton.webp'),
-    proOnly: true,
-  },
-  {
-    name: WireType[WireType.STAGGERED_FUSED_CLAPTON],
-    src: require('@/assets/coil_type_staggered_fused_clapton.webp'),
-    proOnly: true,
-  },
-  {
-    name: WireType[WireType.STAPLE_STAGGERED_FUSED_CLAPTON],
-    src: require('@/assets/coil_type_staple_staggered_fused_clapton.webp'),
-    proOnly: true,
-  },
-  {
-    name: WireType[WireType.FRAMED_STAPLE],
-    src: require('@/assets/coil_type_juggernaut.webp'),
-    proOnly: true,
-  },
-  {
-    name: WireType[WireType.CUSTOM],
-    src: require('@/assets/coil_type_custom.webp'),
-    proOnly: true,
-  },
-];
 
 const ComplexWire: React.FC<WireComponentProps> = props => {
   const {
